backend/requirements.txt
flask
flask-cors
mysql-connector-python
openai==0.28
python-dotenv
